refactor: move drag&drop logic to instanceScreen
