+++ src/main/kotlin/com/mineinabyss/launchy/ui/screens/home/HomeScreen.kt
@@ -29,9 +29,6 @@ fun HomeScreen() {
     Scaffold { paddingValues ->
         val scrollState = rememberLazyListState()
         BoxWithConstraints {
-            val showDragFileCard = remember { mutableStateOf(false) }
-            HandleFileDropping(showDragFileCard)
-            ShowDropFileCard(showDragFileCard)
             Column(Modifier.padding(end = 20.dp).fillMaxSize()) {
 //                var searchQuery by remember { mutableStateOf("") }
 //                SearchBar(
@@ -72,60 +69,6 @@ fun HomeScreen() {
                         hoverColor = MaterialTheme.colorScheme.primary
                     )
                 )
-            }
-        }
-    }
-}
-
-@Composable
-private fun HandleFileDropping(showDragFileCard: MutableState<Boolean>) {
-    val state = LocalLaunchyState
-    val target = object : DropTarget() {
-        @Synchronized
-        override fun dragEnter(event: DropTargetDragEvent) {
-            val files = (event.transferable.getTransferData(DataFlavor.javaFileListFlavor) as List<*>).filterIsInstance<File>()
-            if (files.any { it.extension == "mrpack" })
-                showDragFileCard.value = true
-            else event.rejectDrag()
-        }
-
-        @Synchronized
-        override fun dragExit(event: DropTargetEvent) {
-            showDragFileCard.value = false
-        }
-
-        @Synchronized
-        override fun drop(event: DropTargetDropEvent) {
-            showDragFileCard.value = false
-            runCatching {
-                event.acceptDrop(DnDConstants.ACTION_REFERENCE)
-                val files = (event.transferable.getTransferData(DataFlavor.javaFileListFlavor) as List<*>).filterIsInstance<File>()
-                files.firstOrNull { it.extension == "mrpack" }?.let {
-                    //TODO This does not work as GameInstanceConfig#read doesnt properly deserialize .mrpack
-                    //GameInstance.create(state, GameInstanceConfig.read(it.toPath()))
-                }
-            }.onFailure {
-                it.printStackTrace()
-            }
-        }
-    }
-    windowScope.window.dropTarget = target
-}
-
-@Composable
-private fun ShowDropFileCard(showDragFileCard: MutableState<Boolean>) {
-    if (!showDragFileCard.value) return
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = MaterialTheme.colorScheme.background.copy(alpha = 0.5f)
-    ) {
-        Box(
-            modifier = Modifier.fillMaxSize(),
-            contentAlignment = Alignment.Center
-        ) {
-            Row {
-                Icon(Icons.Rounded.FileUpload, contentDescription = "Upload")
-                Text("Drop a .mrpack to create an instance...", style = MaterialTheme.typography.bodyLarge)
             }
         }
     }
